refactoring UI state with useContext context file created
